Almost working
SO CLOSE! Literally close off the final room and then the only issue is the recursion of allDirectionsTried which is still a bit beyond me. If it iterates over and over, instead of deleting whole map, make it go back to backtracking position 3 (e.g. the second room) as this will hopefully have a direction free. keep going up/down list until finds a free room to branch off.